renamed offerHelp to offer-help
