+++ ClientApp/src/app/areas/offer-help/offer-help-area.pluggable.ts
@@ -1,0 +1,8 @@
+export const offerHelpAreaPluggable = {
+  lazyRoutes: [
+    {
+      path: 'offer-help',
+      loadChildren: () => import('./offer-help-area.module').then((m) => m.OfferHelpAreaModule),
+    },
+  ],
+};
