+++ ClientApp/src/app/areas/offer-help/offer-help-page/offer-help-page.component.html
@@ -1,12 +1,8 @@
-<div class="offerhelp-page-container">
+<div class="offer-help-page-container">
   <ng-container *ngIf="state | async as state">
     <div class="container" [ngSwitch]="state">
       <section class="section" *ngSwitchCase="states.Unknown">
         <p>Loading State</p>
-      </section>
-
-      <section class="section" *ngSwitchCase="states.Complete">
-        <p>Complete State</p>
       </section>
     </div>
   </ng-container>
