+++ ClientApp/src/app/areas/offer-help/offer-help-page/offer-help-page.component.scss
