+++ ClientApp/src/app/areas/offer-help/offer-help-page/orders.service.ts
@@ -1,0 +1,12 @@
+import {Injectable} from '@angular/core';
+import {OrdersGateway} from '../../../core/backend-api/gateways/orders.gateway';
+
+@Injectable({
+  providedIn: 'root',
+})
+export class OrdersService {
+  constructor(
+    private readonly ordersGateway: OrdersGateway,
+  ) {
+  }
+}
