--- ClientApp/src/app/areas/offerhelp/offerhelp-area.pluggable.ts
@@ -1,8 +1,0 @@
-export const offerHelpAreaPluggable = {
-  lazyRoutes: [
-    {
-      path: 'offerhelp',
-      loadChildren: () => import('./offerhelp-area.module').then((m) => m.OfferHelpAreaModule),
-    },
-  ],
-};
