--- ClientApp/src/app/areas/offerhelp/order.model.ts
@@ -1,13 +1,0 @@
-export interface Order {
-  id?: number;
-  orderItems: OrderItem[]
-}
-
-export interface OrderItem {
-  id?: number;
-  product: string;
-  items: number;
-  maxPricePerItem: number;
-  comment: string;
-}
-
